Change visual to audiovisuals in UI for the end game.

+++ LDJAM55/Assets/Scripts/ResultsUI.cs
@@ -19,13 +19,13 @@ public class ResultsUI : MonoBehaviour
         DataTypes.FinalResult finelResult = backend.ProductState.ComputeFinalResult();
 
         // TODO: Add special cases for 1st and 2nd and 3rd
-        overallText.text = "Overall " + ((int)finelResult.overallRank).ToString() + "th";
-        funText.text = "Fun " + ((int)finelResult.funRank).ToString() + "th";
-        innovationText.text = "Innovation " + ((int)finelResult.innovationRank).ToString() + "th";
-        themeText.text = "Theme " + ((int)finelResult.themeRank).ToString() + "th";
-        graphicsText.text = "Graphics " + ((int)finelResult.graphicsRank).ToString() + "th";
-        audioText.text = "Audio " + ((int)finelResult.audioRank).ToString() + "th";
-        humorText.text = "Humor " + ((int)finelResult.humorRank).ToString() + "th";
-        moodText.text = "Mood " + ((int)finelResult.moodRank).ToString() + "th";
+        overallText.text = "Overall " + (finelResult.overallRank).ToString() + "th";
+        funText.text = "Fun " + (finelResult.funRank).ToString() + "th";
+        innovationText.text = "Innovation " + (finelResult.innovationRank).ToString() + "th";
+        themeText.text = "Theme " + (finelResult.themeRank).ToString() + "th";
+        graphicsText.text = "Graphics " + (finelResult.graphicsRank).ToString() + "th";
+        audioText.text = "Audio " + (finelResult.audioRank).ToString() + "th";
+        humorText.text = "Humor " + (finelResult.humorRank).ToString() + "th";
+        moodText.text = "Mood " + (finelResult.moodRank).ToString() + "th";
     }
 }
